+++ src/Celeste.jl
@@ -10,6 +10,7 @@ import ModelInit
 import SampleData
 import Transform
 import KL
+import CelesteCluster
 
 using CelesteTypes
